# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateSigningRequestDocumentData = { [key: string]: unknown };

export interface CreateSigningRequest {
  documentType: string;
  documentTitle: string;
  documentData: CreateSigningRequestDocumentData;
  expiresInDays?: number;
}

export interface SigningRequestCreated {
  token: string;
  signingUrl: string;
  expiresAt: string;
}

export type SigningRequestPublicDocumentData = { [key: string]: unknown };

export interface SigningRequestPublic {
  token: string;
  documentType: string;
  documentTitle: string;
  documentData: SigningRequestPublicDocumentData;
  status: string;
  /** @nullable */
  signerName?: string | null;
  /** @nullable */
  signedAt?: string | null;
  expiresAt: string;
}

export interface SubmitSignature {
  signerName: string;
  signatureData: string;
}