# lib/api-zod/src/generated/types/createSigningRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSigningRequestDocumentData } from "./createSigningRequestDocumentData";

export interface CreateSigningRequest {
  documentType: string;
  documentTitle: string;
  documentData: CreateSigningRequestDocumentData;
  expiresInDays?: number;
}