# lib/api-zod/src/generated/types/signingRequestPublic.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SigningRequestPublicDocumentData } from "./signingRequestPublicDocumentData";

export interface SigningRequestPublic {
  token: string;
  documentType: string;
  documentTitle: string;
  documentData: SigningRequestPublicDocumentData;
  status: string;
  /** @nullable */
  signerName?: string | null;
  /** @nullable */
  signedAt?: string | null;
  expiresAt: string;
}